// packages/ui/app/_grpc/bounds_pb.ts
// @generated by protoc-gen-es v2.5.1 with parameter "target=ts"
// @generated from file bounds.proto (package pb, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Coordinate } from "./coordinate_pb";
import { file_coordinate } from "./coordinate_pb";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file bounds.proto.
 */
export const file_bounds: GenFile = /*@__PURE__*/
  fileDesc("Cgxib3VuZHMucHJvdG8SAnBiIlAKBkJvdW5kcxIgCgh0b3BfbGVmdBgBIAEoCzIOLnBiLkNvb3JkaW5hdGUSJAoMYm90dG9tX3JpZ2h0GAIgASgLMg4ucGIuQ29vcmRpbmF0ZUIuWixnaXRodWIuY29tL2dyZWdoYXJ0L2NsaW1iaW5nLWFwcC9pbnRlcm5hbC9wYmIGcHJvdG8z", [file_coordinate]);

/**
 * @generated from message pb.Bounds
 */
export type Bounds = Message<"pb.Bounds"> & {
  /**
   * @generated from field: pb.Coordinate top_left = 1;
   */
  topLeft?: Coordinate;

  /**
   * @generated from field: pb.Coordinate bottom_right = 2;
   */
  bottomRight?: Coordinate;
};

/**
 * Describes the message pb.Bounds.
 * Use `create(BoundsSchema)` to create a new message.
 */
export const BoundsSchema: GenMessage<Bounds> = /*@__PURE__*/
  messageDesc(file_bounds, 0);