// packages/ui/app/_grpc/commentable_pb.ts
// @generated by protoc-gen-es v2.5.1 with parameter "target=ts"
// @generated from file commentable.proto (package pb, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Comment } from "./comment_pb";
import { file_comment } from "./comment_pb";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file commentable.proto.
 */
export const file_commentable: GenFile = /*@__PURE__*/
  fileDesc("ChFjb21tZW50YWJsZS5wcm90bxICcGIiTAoLQ29tbWVudGFibGUSCgoCaWQYASABKAMSEgoKZGVzY3JpcHRvchgCIAEoCRIdCghjb21tZW50cxgDIAMoCzILLnBiLkNvbW1lbnRCLlosZ2l0aHViLmNvbS9ncmVnaGFydC9jbGltYmluZy1hcHAvaW50ZXJuYWwvcGJiBnByb3RvMw", [file_comment]);

/**
 * @generated from message pb.Commentable
 */
export type Commentable = Message<"pb.Commentable"> & {
  /**
   * @generated from field: int64 id = 1;
   */
  id: bigint;

  /**
   * @generated from field: string descriptor = 2;
   */
  descriptor: string;

  /**
   * @generated from field: repeated pb.Comment comments = 3;
   */
  comments: Comment[];
};

/**
 * Describes the message pb.Commentable.
 * Use `create(CommentableSchema)` to create a new message.
 */
export const CommentableSchema: GenMessage<Commentable> = /*@__PURE__*/
  messageDesc(file_commentable, 0);