// packages/ui/app/_grpc/coordinate_pb.ts
// @generated by protoc-gen-es v2.5.1 with parameter "target=ts"
// @generated from file coordinate.proto (package pb, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file coordinate.proto.
 */
export const file_coordinate: GenFile = /*@__PURE__*/
  fileDesc("ChBjb29yZGluYXRlLnByb3RvEgJwYiImCgpDb29yZGluYXRlEgsKA2xhdBgBIAEoARILCgNsbmcYAiABKAFCLlosZ2l0aHViLmNvbS9ncmVnaGFydC9jbGltYmluZy1hcHAvaW50ZXJuYWwvcGJiBnByb3RvMw");

/**
 * @generated from message pb.Coordinate
 */
export type Coordinate = Message<"pb.Coordinate"> & {
  /**
   * @generated from field: double lat = 1;
   */
  lat: number;

  /**
   * @generated from field: double lng = 2;
   */
  lng: number;
};

/**
 * Describes the message pb.Coordinate.
 * Use `create(CoordinateSchema)` to create a new message.
 */
export const CoordinateSchema: GenMessage<Coordinate> = /*@__PURE__*/
  messageDesc(file_coordinate, 0);