// packages/ui/app/_grpc/photoable_pb.ts
// @generated by protoc-gen-es v2.5.1 with parameter "target=ts"
// @generated from file photoable.proto (package pb, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file photoable.proto.
 */
export const file_photoable: GenFile = /*@__PURE__*/
  fileDesc("Cg9waG90b2FibGUucHJvdG8SAnBiIisKCVBob3RvYWJsZRIKCgJpZBgBIAEoAxISCgpkZXNjcmlwdG9yGAIgASgJQi5aLGdpdGh1Yi5jb20vZ3JlZ2hhcnQvY2xpbWJpbmctYXBwL2ludGVybmFsL3BiYgZwcm90bzM");

/**
 * @generated from message pb.Photoable
 */
export type Photoable = Message<"pb.Photoable"> & {
  /**
   * @generated from field: int64 id = 1;
   */
  id: bigint;

  /**
   * Crag crag = 3; // Omitted to avoid circular reference
   * Area area = 4; // Omitted to avoid circular reference
   * Boulder boulder = 5; // Omitted to avoid circular reference
   * Route route = 6; // Omitted to avoid circular reference
   *
   * @generated from field: string descriptor = 2;
   */
  descriptor: string;
};

/**
 * Describes the message pb.Photoable.
 * Use `create(PhotoableSchema)` to create a new message.
 */
export const PhotoableSchema: GenMessage<Photoable> = /*@__PURE__*/
  messageDesc(file_photoable, 0);